Add drugs seeders, fix users in user conditions seeder

+++ hf-app/database/migrations/2021_02_23_203320_create_drugs_table.php
@@ -16,7 +16,7 @@ class CreateDrugsTable extends Migration
         Schema::create('drugs', function (Blueprint $table) {
             $table->id();
             $table->string('name')->unique();
-            $table->integer('dosage_volume')->nullable();
+            $table->float('dosage_volume')->nullable();
             $table->string('dosage_unit')->nullable();
             $table->integer('dosage_times')->nullable();
             $table->enum('dosage_span', ['hour', 'day', 'week', 'month'])->nullable();
